backend_python_project/requirements.txt
fitz
google.generativeai
io
base64
flask_cors
PIL
os
dotenv
flask